+++ src/simplex.rs
@@ -14,6 +14,7 @@
 // along with MGF. If not, see <http://www.gnu.org/licenses/>.
 
 use std::f32;
+use std::fmt;
 
 use cgmath::prelude::*;
 use cgmath::{EuclideanSpace, InnerSpace, Point3, Vector3};
@@ -24,6 +25,34 @@ pub struct Simplex {
     state: &'static SimplexState,
 }
 
+impl fmt::Debug for Simplex {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        #[derive(Debug)]
+        enum StateName {
+            Vertex,
+            Edge,
+            Face,
+            Volume
+        };
+        let (state, points) = if self.state as *const _ == &VERTEX_DATAPTRLOC as *const _ {
+            (StateName::Vertex, &self.points[..1])
+        } else if self.state as *const _ == &EDGE_DATAPTRLOC as * const _  {
+            (StateName::Edge, &self.points[..2])
+        } else if self.state as * const _  == &FACE_DATAPTRLOC as * const _ {
+            (StateName::Face, &self.points[..3])
+        } else {
+            (StateName::Volume, &self.points[..4])
+        };
+        write!(f, "Simplex {{ state: {:?}, points: {:?} }}", state, points)
+    }
+}
+
+impl From<[Point3<f32>; 1]> for Simplex {
+    fn from(p: [Point3<f32>; 1]) -> Self {
+        Self::from(p[0])
+    }
+}
+          
 impl From<Point3<f32>> for Simplex {
     fn from(p: Point3<f32>) -> Self {
         Simplex {
@@ -35,6 +64,12 @@ impl From<Point3<f32>> for Simplex {
             ],
             state: &VERTEX_DATAPTRLOC,
         }
+    }
+}
+
+impl From<[Point3<f32>; 2]> for Simplex {
+    fn from(p: [Point3<f32>; 2]) -> Self {
+        Self::from((p[0], p[1]))
     }
 }
 
@@ -52,6 +87,12 @@ impl From<(Point3<f32>, Point3<f32>)> for Simplex {
     }
 }
 
+impl From<[Point3<f32>; 3]> for Simplex {
+    fn from(p: [Point3<f32>; 3]) -> Self {
+        Self::from((p[0], p[1], p[2]))
+    }
+}
+          
 impl From<(Point3<f32>, Point3<f32>, Point3<f32>)> for Simplex {
     fn from(p: (Point3<f32>, Point3<f32>, Point3<f32>)) -> Self {
         Simplex {
@@ -63,6 +104,12 @@ impl From<(Point3<f32>, Point3<f32>, Point3<f32>)> for Simplex {
             ],
             state: &FACE_DATAPTRLOC,
         }
+    }
+}
+
+impl From<[Point3<f32>; 4]> for Simplex {
+    fn from(p: [Point3<f32>; 4]) -> Self {
+        Self::from((p[0], p[1], p[2], p[3]))
     }
 }
 
@@ -80,35 +127,24 @@ impl From<(Point3<f32>, Point3<f32>, Point3<f32>, Point3<f32>)> for Simplex {
     }
 }
 
-pub enum GJKIterResult {
-    Incomplete,
-    Separation(f32),
-    Penetration(f32),
-}
-
 impl Simplex {
-    /// Performs a single iteration of GJK
-    pub fn iter_gjk<AT, BT>(&mut self, a: &AT, b: &BT) -> GJKIterResult
+    /// Finds the closest point on the shape to the origin
+    pub fn closest_point_to_origin<S>(&mut self, shape: &S) -> Point3<f32>
     where
-        AT: Convex,
-        BT: Convex
+        S: Convex
     {
-        let (min_norm, next_state) = self.state.min_norm(&mut self.points);
-        self.state = next_state;
-        if min_norm.magnitude2() <= COLLISION_EPSILON {
-            return GJKIterResult::Penetration(
-                // TODO: Calculate this, somehow
-                0.0
-            );
+        loop {
+            let (min_norm, next_state) = self.state.min_norm(&mut self.points);
+            self.state = next_state;
+            if min_norm.magnitude2() < COLLISION_EPSILON {
+                return Point3::from_vec(min_norm);
+            }
+            let support = shape.support(-min_norm.normalize()).to_vec();
+            if min_norm.magnitude2() == support.magnitude2() {
+                return Point3::from_vec(min_norm);
+            }
+            self.state.add_point(&mut self.points, support);
         }
-        let v = a.support(-min_norm) - b.support(min_norm);
-        if v.dot(min_norm) <= 0.0 {
-            return GJKIterResult::Separation(
-                min_norm.magnitude()
-            )
-        }
-        self.state.add_point(&mut self.points, v);
-        GJKIterResult::Incomplete
     }
 }
 
@@ -165,14 +201,16 @@ impl SimplexState for FaceSimplex {
     fn min_norm(&self, simp: &mut [Vector3<f32>; 4]) -> (Vector3<f32>, &'static SimplexState) {
         let ab = simp[1] - simp[0];
         let ac = simp[2] - simp[0];
-
         let ap = -simp[0];
         let d1 = ab.dot(ab);
         let d2 = ac.dot(ap);
+
+        // Vertex region A
         if d1 <= 0.0 && d2 <= 0.0 {
             return (simp[0], &EDGE_DATAPTRLOC);
         }
 
+        // Vertex region B
         let bp = -simp[1];
         let d3 = ab.dot(bp);
         let d4 = ac.dot(bp);
@@ -181,12 +219,14 @@ impl SimplexState for FaceSimplex {
             return (simp[1], &EDGE_DATAPTRLOC);
         }
 
+        // Edge region AB
         let vc = d1 * d4 - d3 * d2;
         if vc <= 0.0 && d1 >= 0.0 && d3 <= 0.0 {
             let v = d1 / (d1 - d3);
             return ((simp[0] + ab * v), &FACE_DATAPTRLOC);
         }
 
+        // Vertex region C
         let cp = -simp[2];
         let d5 = ab.dot(cp);
         let d6 = ac.dot(cp);
@@ -195,6 +235,7 @@ impl SimplexState for FaceSimplex {
             return (simp[2], &EDGE_DATAPTRLOC);
         }
 
+        // Edge region AC
         let vb = d5 * d2 - d1 * d6;
         if vb <= 0.0 && d2 >= 0.0 && d6 <= 0.0 {
             let w = d2 / (d2 - d6);
@@ -202,6 +243,7 @@ impl SimplexState for FaceSimplex {
             return ((simp[0] + ac * w), &FACE_DATAPTRLOC);
         }
 
+        // Edge region BC
         let va = d3 * d6 - d5 * d4;
         if va <= 0.0 && (d4 - d3) >= 0.0 && (d5 - d6) >= 0.0 {
             let w = (d4 - d3) / ((d4 - d3) + (d5 - d6));
@@ -237,12 +279,14 @@ impl SimplexState for VolumeSimplex {
         let (a, b, c, d) = (simp[0], simp[1], simp[2], simp[3]);
         // Test face abc
         if origin_outside_plane(a, b, c, d) {
-            let (p, new_state) = FACE_DATAPTRLOC.min_norm(simp);
+            let mut new_simp = [ a, b, c, d ];
+            let (p, new_state) = FACE_DATAPTRLOC.min_norm(&mut new_simp);
             let new_dist = p.magnitude2();
             if new_dist < best_dist {
                 closest_pt = p;
                 best_dist = new_dist;
                 next_state = new_state;
+                *simp = new_simp;
             }
         }
         // Test face acd